Restructure as a set of ES modules
Help with providing similar features in the browser
