+++ lib/octokit.mjs
@@ -2,13 +2,13 @@
  * Wrapper around Octokit to add throttling and avoid hitting rate limits
  * TODO: DRY with Webref?
  */
-
-const { throttling } = require("@octokit/plugin-throttling");
-const Octokit = require("@octokit/rest").Octokit.plugin(throttling);
+import { throttling } from "@octokit/plugin-throttling";
+import OctokitRest from "@octokit/rest";
+const DefaultOctokit = OctokitRest.Octokit.plugin(throttling);
 
 const MAX_RETRIES = 3;
 
-module.exports = function (params) {
+export function Octokit (params) {
   params = params || {};
 
   const octoParams = Object.assign({
@@ -34,5 +34,5 @@ module.exports = function (params) {
     }
   }, params);
 
-  return new Octokit(octoParams);
+  return new DefaultOctokit(octoParams);
 }
