+++ lib/parse-minutes.mjs
@@ -1,11 +1,4 @@
-const { JSDOM } = require("jsdom");
-const Octokit = require("./lib/octokit");
-
-let dryRun = false;
-
-let GH_TOKEN, octokit;
-
-const githubRe = /^https:\/\/github.com\/([^\/]+)\/([^\/]+)\/(issues|pull)\/([0-9]+)/;
+export const githubRe = /^https:\/\/github.com\/([^\/]+)\/([^\/]+)\/(issues|pull)\/([0-9]+)/;
 const headingSelector = ":is(h1,h2,h3,h4,h5,h6";
 
 // Find the closest previous heading associated to the given element
@@ -20,16 +13,14 @@ function getCurrentHeading(el) {
   return;
 }
 
-async function parseMinutes(url) {
+
+export async function parseMinutes(document, url) {
   const foundLinks = new Set();
   // annotatedLinks can be set with a type of:
   // resolved: issue/PR linked from a resolution or from topic or subtopic with a resolution
   // discussed: issue/PR linked from a topic or subtopic in the minutes
   // mentioned: issue/PR linked from the minutes
   const annotatedLinks = [];
-
-  const dom = await JSDOM.fromURL(url);
-  const document = dom.window.document;
 
   const title = document.querySelector("title").textContent;
 
@@ -91,73 +82,3 @@ async function parseMinutes(url) {
   return {url, title, annotatedLinks};
 }
 
-function formatGithubComment(annotatedLink, url, title) {
-  let state;
-  switch(annotatedLink.type) {
-  case "resolved":
-    state = "had an associated resolution";
-    break;
-  case "discussed":
-    state = "was discussed";
-    break;
-  default:
-    state = "was mentioned";
-  }
-  let comment = `This issue ${state} in [the minutes of ${title} ${annotatedLink.context ? ` (${annotatedLink.context.title})`: ""}](${url}#${annotatedLink?.context?.id || ''})`;
-  if (annotatedLink.type === "resolved") {
-    comment += `:
-
-> ${annotatedLink.resolution}
-`;
-  }
-  return comment;
-}
-
-async function updateGithub({url, title, annotatedLinks}) {
-  for (const annotatedLink of annotatedLinks) {
-    const comment = formatGithubComment(annotatedLink, url, title);
-    // submit on relevant github issue
-    const [, owner, repo,, issue_number] = annotatedLink.link.match(githubRe);
-    console.log(`submitting to https://github.com/${owner}/${repo}/issues/${issue_number}`);
-    console.log(comment);
-    if (dryRun) {
-      console.log("- not really, dry run");
-    } else {
-      await octokit.rest.issues.createComment(
-	{owner, repo, issue_number, body: comment}
-      );
-    }
-  }
-}
-
-if (require.main === module) {
-  let minutesUrl;
-  if (process.argv[2] === "--dry-run") {
-    dryRun = true;
-    minutesUrl = process.argv[3];
-  } else {
-    GH_TOKEN = (() => {
-      try {
-	return require("./config.json").GH_TOKEN;
-      } catch (err) {
-	return process.env.GH_TOKEN;
-      }
-    })();
-    if (!GH_TOKEN) {
-      console.error("No github token set, can't submit comments");
-      process.exit(2);
-    }
-    octokit = new Octokit({
-      auth: GH_TOKEN,
-      //log: console
-    });
-
-    minutesUrl = process.argv[2];
-  }
-  parseMinutes(minutesUrl)
-    .then(updateGithub)
-    .catch(err => {
-      console.error(err);
-      process.exit(2);
-    });
-}
